refactor(topology): merge LANG01 and VIZ01-dashboard into parent as feature modules
lang-rust (with optional syn, quote, proc-macro2, walkdir) and viz-dashboard (with optional chrono) complete the five-module merge. aps-cli viz call sites now route through code_topology::substandards::viz_dashboard. New optional external deps added to the standard.toml allowlist with ADR-0002 rationales.
