+++ Cargo.toml
@@ -11,8 +11,6 @@ members = [
     "standards/v1/APS-V1-0000-meta/substandards/DI01-distribution",
     # Official standards
     "standards/v1/APS-V1-0001-code-topology",
-    "standards/v1/APS-V1-0001-code-topology/substandards/VIZ01-dashboard",
-    "standards/v1/APS-V1-0001-code-topology/substandards/LANG01-rust",
     # Experimental standards
     "standards-experimental/v1/EXP-V1-0002-todo-tracker",
     "standards-experimental/v1/EXP-V1-0003-fitness-functions",
