+++ standards/v1/APS-V1-0001-code-topology/src/substandards/mod.rs
@@ -8,3 +8,9 @@ pub mod viz_mermaid;
 
 #[cfg(feature = "viz-3d")]
 pub mod viz_3d;
+
+#[cfg(feature = "lang-rust")]
+pub mod lang_rust;
+
+#[cfg(feature = "viz-dashboard")]
+pub mod viz_dashboard;
