+++ standards/v1/APS-V1-0001-code-topology/src/substandards/viz_dashboard/index.rs
@@ -2,7 +2,7 @@
 //!
 //! Landing page linking to all visualization types with summary statistics.
 
-use crate::{health_label, health_to_color};
+use super::{health_label, health_to_color};
 
 /// Generate dashboard index HTML.
 ///
