--- standards/v1/APS-V1-0001-code-topology/substandards/LANG01-rust/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "apss-v1-0001-lang01-rust"
-version = "0.1.0"
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-description = "Rust language adapter for Code Topology  -  analyzes Rust code and generates .topology/ artifacts"
-
-[lib]
-name = "code_topology_rust_adapter"
-
-
-[dependencies]
-apss-core.workspace = true
-code-topology = { package = "apss-v1-0001-code-topology", path = "../.." }
-serde.workspace = true
-serde_json.workspace = true
-toml.workspace = true
-walkdir.workspace = true
-
-# Rust parsing
-syn = { version = "2", features = ["full", "parsing", "visit", "extra-traits"] }
-quote = "1"
-proc-macro2 = "1"
-
-# Error handling
-thiserror.workspace = true
-
-[dev-dependencies]
-tempfile = "3.14"
-
