--- standards/v1/APS-V1-0001-code-topology/substandards/VIZ01-dashboard/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "apss-v1-0001-viz01-dashboard"
-version = "0.1.0"
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-description = "Topology Visualization Dashboard - Interactive HTML visualizations for code topology"
-
-[lib]
-name = "code_topology_viz"
-
-
-[dependencies]
-toml.workspace = true
-apss-core.workspace = true
-code-topology = { package = "apss-v1-0001-code-topology", path = "../.." }
-serde.workspace = true
-serde_json.workspace = true
-
-# Timestamps for dashboard generation
-chrono = { version = "0.4", features = ["std"] }
-
-[dev-dependencies]
-tempfile = "3.14"
